// frontend/src/app/models/article.ts
export class Article {
    id: string;
    companyUsername: string;
    name: string;
    unitOfMeasure: string;
    taxRate: number;
    type: string;
    origin: string;
    originalName: string;
    barcode: string;
    manufacturer: string;
    customsRate: string;
    ecoTax: boolean;
    exciseTax: boolean;
    desiredStockMin: number;
    desiredStockMax: number;
    about: string;
    declaration: string;
    image: string;
    category: string;
    subcategory: string;
}